+++ main.go
@@ -1,19 +1,43 @@
 package main
 
 import (
+	"errors"
 	"fmt"
 	"os"
+	"time"
 
 	"github.com/sirupsen/logrus"
 )
 
 func main() {
 	logrus.Infof("Starting up")
-	if len(os.Args) > 1 {
-		fmt.Println(os.Args[1])
-	} else {
-		fmt.Println("Hello, I am Gopher!")
+
+	hourOfDay := time.Now().Hour()
+	greeting, err := getGreeting(hourOfDay)
+	if err != nil {
+		fmt.Println(err)
+		os.Exit(1)
 	}
 
+	fmt.Println(greeting)
 	logrus.Infof("Normal exit")
+}
+
+func getGreeting(hour int) (string, error) {
+	var message string
+
+	if hour < 7 {
+		err := errors.New("Too early for Gophers")
+		return message, err
+	}
+
+	if hour < 12 {
+		message = "Good morning"
+	} else if hour < 18 {
+		message = "Good afternoon"
+	} else {
+		message = "Good evening"
+	}
+
+	return message, nil
 }