+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 
 # Unreleased
 
-## 0.0.1 2020-08-18
+- Use a variable to hold the Arguments.
+- Create a function to return a greeting message dependant on the hour of the day.
+- Added an error return value "Too early for Gophers"
+
+## v0.0.1 2020-08-18
 
 - Initial files
 - Conditionals 1: